Show more examples
Change the used context and observe how much is rebuilt. This should be
proof enough that it is the BuildContext that defines how much of the UI
is rebult, not where it is used or how deep syntactically the dependent
widget is.

NOTE: At least a hot restart is sometimes necessary to observe correct
behavior. This manifests itself when changing from a broader context to
an norrower one - the logs still show the outer context being built and
the code seems not to be working as expected. The reason is that hot
reload doesn't reset context dependencies so the context resigered as a
dependency in the previous build is still the one rebuilt after hot
reload.