Demonstrate a 'minor' issue
PageView with 'unexpected' build() calls which unnecessarily trigger
Cubit actions, like HTTP requests.

build() can be called when not expected and developers should never
assume the calls won't happen. It is safest to play by the rules and
make sure build() has no side effects, this way it will always work.

(Note that sometimes rebuilding and reinitializing on each page change
may be desired, e.g. when the data changes often, but in this case the
Cubit probably shouldn't be global but rather scoped to the single page
anyway.)