Demonstrate a 'major' issue
A custom animation which calls build() on each frame.

When build() is called repeatedly, e.g. in a custom animation, this
causes all side effects to be executed over and over again. Not only
this, any potential work resulting from it is also executed
unnecessarily, potentially without any option of cancelling it, causing
unnecessary work for the device and others, e.g. a REST API server.

In certain cases, like HTTP requests, ordering is not guaranteed so the
repsonses will be processes out of order (as they arrive), possibly
wreaking havoc in application state.

NOTE: There are other ways to scale a container which are easier and
perform better; AnimatedBuilder was chosen to clearly present the issue.
Also, it is kind of a proof - using certain code doesn't manifest any
issues but a small change, like in how the animation is done, may blow
up in our faces. No side effect, no issues no matter what.