Demonstrate a 'major' issue visible to end users
Havoc in application state - flickering, enjoy the disco lights.

The widget that is part of the animated UI now depends on the state of
the Cubit that is initialized on every frame, which causes a lot of
flickering due to unnecessary and uncontrolled state changes from
loading to data to loading to other data etc.