Solution #3 - initialization in State.initState
This has the drawback that if the widget is used multiple times in a
page the Cubit would be initialized multiple times as well. The same
happens when different widgets use the same Cubit but all of them
initialize it in their initState(). In such situations the solution
would be to 'lift the state up' and introduce a common parent
StatefulWidget which would initialize the Cubit once.