Prepare demo app
- A Cubit generating Colors.
- A deep tree of widgets reacting to Cubit state changes.

Even though we ignore the state and the widgets don't change when the
Cubit emits new states, all of them are rebuilt anyway.

This is because for Flutter each state change results in a new widget
tree - the widgets are equivalent to us humans (we understand they have
the same type and the same configuration), but to Flutter they are *not
equal* so it needs to assume things changed and need to be rebuilt.