Fix the issue using BlocBuilder
The only thing that depends on the Cubit is now the bottom-most widget,
the other ones don't change at all so they are not rebuilt. They are
still not const but it doesn't matter in this case as only the part
depending on the color is rebuilt by Flutter (but remember to use helper
widgets like here, or to use the right context - see the previous branch
about optimization with BuildContext).