Use const for the whole tree
It is possible to use const here but it doesn't change anything as the
upper-level widgets aren't rebuilt anyway. However, in other scenarios
it might be necessary to be able to squeeze out every last drop of
performance so it is a viable option.

Note that the code got a bit uglier and not as fluently readable as with
the inline closure, so arguably there are drawbacks. The callback needs
to be a static method or a top level function.