Cache a static subtree if const is not available
If a Widget (e.g. Container doesn't have a const constructor) can't be
made static for whatever reason, we ca still make sure Flutter
understands we use the same instance over and over again - we can create
it once, cache and use in build().

This makes the code a bit less readable and isn't often needed, but if
we need all the performance we can get it is a viable option. It is also
a great option when constructing the widget is expensive as it will only
ever be created once.

(Note that creating the child in the build() method before the
BlocBuilder would also work due to the closure that builder callback
creates.)