Limit what is rebuilt using const
Now the static widgets are not rebuilt. The optimization is trivial but
according to the Flutter frames chart from DevTools average frame time
drops by half (although the suboptimal case without const is still so
fast that no jank is possible due to the UI being so simple - at least
on my device ;D).