Limit what is rebuilt using AnimatedBuilder.child
The static widgets are not rebuilt either.

It is possible to build the static part of the tree once and give it
over to AnimatedBuilder, which will cache it internally and pass to the
builder callback. This will make Flutter understand not to rebuild this
subtree as the instances are the same. Note that no const is involved
this time around so this child caching is a very good option when const
is not possible.

A few examples of other widgets whose builder callbacks support the
child parameter:
- ValueListenableBuilder (Flutter).
- TweenAnimationBuilder (Flutter).
- Consumer (Provider package).
- Consumer (Riverpod package).