Enter InheritedWidget
Now the count is provided at the very top of the tree and read only
where necessary, which has the following advantages:

- All intermediate widgets are blissfully oblivious to the data that
  they are not interested in.
- We can use const for these widgets as they don't depend on runtime
  values.
- The code is cleaner and we will be able to handle a lot of data this
  way.

Disadvantages:
- Transferring the data is not explicit.
- If we forget to provide the data, the app will fail at runtime.
  Experience says this commonly happens in widget tests.
- It is only possible to use the same InheritedWidget only once per tree
  as the lookup is done by type.

(This is the way things have been traditionally done in Flutter, but
there are libraries trying to remedy these issues, e.g. Riverpod.)

We can get the count, but we can't update it. InheritedWidget and its
subclasses are immutable so to set a new count we somehow need to get
the build() method creating the InheritedWidget executed and return a
new instance - this way Flutter will discover the change.