Implement (broken) inherited count updates
We need to do something that causes the InheritedWidget subclass to be
rebuilt. Very often this is done using a StatefulWidget wrapper, e.g.
Localizations and Theme, and this is what we do as well.

We look for State in increment() because this is now our source of truth
and keeps the logic related to count - corresponds to what Navigator is
doing. InheritedCount is now just a 'window' to the state passed down
the UI tree.

Still, MyWidget is not rebuilt with the new count. When we cause hot
reload, though, we see that the value has changed - the reason is that
our current code does change the state/count, but it doesn't trigger
MyWidget to rebuild; hot reload rebuilds the whole app but reuses the
state so the effect is visible.