Fix by registering dependent widgets correctly
Flutter will now understand that users of the InheritedCount.of() method
are dependent wigets and want to be notified of changes and rebuilt.

This works by widgets* registering themselves as dependants on
InheritedWidgets. When InheritedWidget changes its updateShouldNotify()
method is called and if it returns true all dependent widgets rebuild.

A widget can have multiple dependencies. In the case of MyWidget there
are 2: InheritedCount and Theme, but there could be more, like
Localizations, MediaQuery, some Cubit etc. A rebuild will be triggered
whenever any of the dependencies change.

*It's actually not Widgets that have dependents and dependencies, its
their underlying Elements, but let's keeps things simple for now.

The way it works internally is simple - each Element has a collection of
dependencies and each InheritedWidget has a collection of dependents.
Also, all Elements have a collection of available InheritedWidgets above
them. When a new Element is introduced, it gets the same collection from
its parent (for quick lookups without having to traverse the tree up, it
gets the same instance to avoid copying), and each time a new
InheritEdelement is introduced into the tree it copies this collection,
adds itself to it, and this will be passed to children (this is done so
that only descendants see the new InheritedElement; if there was no copy
made, adding would mutate the collection for all parents as well - not
good).

When an Element registers itself as a dependency on an InheritedElement,
it is simply added to the dependents of that element, and that element
is also added as a dependency of this element. When a new
InheritedWidget is detected, the InheritedElement behind it checks if
the change is meaningful (calls updateShouldNotify()) and if yes, it
iterates over its dependencies and marks them for rebuild in the next
frame.