Demonstrate when filtering update notifications
FAB long press now sets the value to a fix number (5). When the current
count is different, we see that the screen rebuilds and everything
works. However, when the current counter value happens to be the value
that we set to, we see that the screen doesn't rebuild (debug console
doesn't show a new entry). This is becaue updateShouldNotify() does its
job.