Change the kind of data InheritedCount keeps
We now return an object which exposes some operations. In reality this
is the state object which implements a custom interface. This is kind
of what Navigator.of() does ( minus the interface, they return
NavigatorState directly).

InheritedWidget data can be arbitrarily complex. For example, a very
common object is Cubit or Bloc - they use the Provder package under the
hood, which is 'A wrapper around InheritedWidget' (quote from the docs
as of now).