Introduce InheritedNotifier for easier mutation
Sometimes, especially when we need to exchange the data the
InheritedWidget provides, using a ValueNotifier with InheritedNotifier
is a good option.

ValueNotifier is a mutable wrapper around a value which allows us to:
- Get the current value.
- Register listener for value change events.
- Set the current value. Whenever the value is set to a different one
  according to operator ==, all listeners are notified.

InheritedNotifier gets a Listenable (ValueNotifier implements
Listenable) and registers itself as its listener. When the value is
changed, InheritedNotifier is notified and in turn notifies dependent
widgets that a rebuild is needed.

So far we've seen 3 different ways of implementing the same thing -
counter updates and reads. There are of course as many (or more) as
there are developers ;D even Flutter implements it in different ways
across the framework.

It is worh mentioning that there also exists InheritedModel which makes
use of the optioal 'aspect' parameter to the
dependOnInheritedWidgetOfExactType() method to support notifications
when a particular part (the aspect) of the state changes - this doesn't
notify widgets that are not interested in this part.