Refactor a bit for a poor man's Provider.
The code with InheritedWidget is in general pretty tricky and there is a
lot of it so it's best not to have to write it too often and reuse.

What we have now is a very rudimentary but functional and generic code
somewhat similar to the popular Provider* package. The read and watch
BuildContext extensions should be familiar as well as the Bloc library
uses Provider internally and exposes them.

*Provider is much more robust and supports special provider types for
Streams, Futures etc. but our code is a good start.