Use Provider package
It is better than ours anyway ;D

Summary: working with InheritedWidget involves some tricky code, calls
to methods with strange names and quite a lot of boilerplate. Normally,
thanks to packages like Provider, Riverpod etc. we don't have to do it
often (even never), but it's good to understand how it works.