Implement a basic builder
It doesn't do anything yet except for printing something out. Also,
there is no way to execute it yet, we need more packages for this.

The build package is the most basic building block for asset generation.
It defines a few key concepts:

Builder: defines the logic for generation. Its build() method will be
called for each input file represented as a BuildStep.

BuildStep: represent each input file for the Builder.build() method.
This means this method will be called as many times as there are input
files. This is also the class used to read and write files, kind of an
abstraction of a filesystem, which allows the package to track such
operations, validate that rules are not violated (e.g. creating a
disallowed file) etc. Builders should never ever use the filesystem APIs
directly.

AssetId: represents an asset of a package. An asset can be anything: a
Dart source file, an image, even the pubspec.yaml file is an asset. Most
commonly this type is used to represent the main input file for a
BuildStep, but it is also possible to look up other assets using
BuildStep APIs.

Resolver: a higher level API over the analyzer package, allowing
operations on Dart source code during a build.

Each builder must define a mapping in its buildExtensions property. It
associates input extensions with one or more output extensions. All
input files that the Builder.build() method will ever be called with are
ones that match the extensions defined as keys in this map. All output
files that the builder is allowed to write are ones that match the
extensions defined as values in this map. Normally it works by a simple
substitution, e.g. for a mapping .dart -> [.g.dart] and a file
lib/customer.dart a file called lib/customer.g.dart can be output. This
is actually very important and strictly enforced by the asset writing
methods mentioned above. Multiple output extensions means multiple files
can be written.

Output suffixes must not overlap for multiple builders, i.e. if one
builder defines a .g.dart output, no other can do it. It's not as bad as
it sounds, there are ways around it.

Builders must not overwrite existing files, only add new ones, the build
will fail otherwise.