Make it possible to run the builder
In order to run builders another package is necessary: build_runner. It
is added as a dev dependency to the builder clients, i.e. packages that
want to run builders. It has a few commands, the most useful ones are
run (for a one-off build) and watch (for continuous rebuilding when
input files keep changing, e.g. during development).

To trigger a build using the build_runner package we can use
flutter pub run build_runner build
for Flutter, or
dart run build_runner build
for pure Dart.
Alternatively, in VS Code shortuts can be used (CMD+SHIFT+B on Mac).

Currently, our build runs but shows a severe warning that there is
nothing to be built - we are missing a configuration build.yaml, which
is hinted at by the --config switch in the help of the build command
(flutter pub run build_runner build --help).