Configure builder for dependent packages
To configure the builder we need a dependency on the build_config
package (actually, it is already a transitive dependency but it should
be a direct dependency - not a dev dependency - so that clients can be
sure that they are consuming a compatible configuration file).

The build.yaml file in the hello_generator package describes the builder
(where it is, how it is created, what it applies to, where outputs are
generated, etc.). More information on the file:
https://github.com/dart-lang/build/tree/master/build_config#defining-builders-to-apply-to-dependents
https://github.com/dart-lang/build/blob/master/docs/build_yaml_format.md

Running the build now shows our message being logged to the console.

NOTE: It will be necessary to get all dependencies in both packages.

The build.yaml file repeats the build extensions configuration map. This
is actually kind of redundant because the builder does the same, but
this map is static and available to the build system without ever
running code, allowing the system to create a builder dependency graph.
There are talks of maybe removing this. In reality, this map in
build.yaml can be different than the one returned buy the builder - it
will not explode but the system might work in an undefined manner. Just
in case, the
flutter pub run build_runner doctor
command can tell us about inconsistencies in configuration.

NOTE: The build system caches build outputs aggresively so if nothing or
just the configuration changes running the build will not execute any
actions. To force this, call
flutter pub run build_runner clean
and then build again - the last line of output should say that 1 action
was executed, but no outputs were generated, which is fine as our
builder just logs as of now.