Add a run configuration for easy builder debugging
The build system uses code generation itself and it generates the build
script in .dart_tool/build/entrypoint/build.dart. When we open the file
we can see our builder configured there with values from build.yaml. The
script is generated as part of the build etc. commands, but we can have
it generated without building as well using
flutter pub run build_runner generate-build-script

NOTE: This command doesn't appear in the help output for build_runner,
and neither does doctor. I stumbled upon both in the depths of the
internet.

As the build script is simply a Dart application that we can call, we
can actually create a run configuration calling the script to be able to
debug our builder.

NOTE: Due to aggressive caching of build results, it is necessary to
often clean the cache with
flutter pub run build_runner clean
unless the builder implementation changes.