Generate an asset
Now our builder generates a .hello file for each input .dart file, right
next to it in the lib folder of the main application package, the client
of the builder.

The builder code shows a few APIs: how to work with the input asset, how
to construct the name of the output and write it.

Change the output extension in the build method to something else than
.hello and watch the build validate the output and fail.