Generate Dart code and use in the app
Builders can automatically apply other builder to run afterwards in the
applies_builders field. Builders can also request that the build system
expose outputs from other builders so that they use them as inputs using
the required_inputs field. All of this allows the build system to order
builders correctly into 'phases' and create a kind of a build pipeline.

We now have 2 builders:
1. hello: Creates .hello.part output files in the cache directory for
   all .dart input files. This is a builder that defines the main public
   interface of the package. The fact that it in reality only generates
   part files and needs help to convert that to Dart code is an
   implementation detail.
2. _hello_dart: Creates .hello.dart output files in the /lib directory
   alongside other .dart files and uses the output of hello builder as
   input. The name starts with a _ to convey that it is a private
   builder not meant for clients to enable or configure explicitly - it
   is applied automatically and is in fact an implementation detail of
   the package, but it still needs to be publicly known for the build
   system to work. json_serializable does something similar.

NOTE: These builders are just an example and in reality a single builder
would be enough to get the same result.

We now define that the generated .dart files should be committed as they
are used in the app, which is what packages like Freezed or
json_serializable do as well.