Collect all parts into a single output
It is possible to use 'synthetic' inputs to act on all files in a
package (others are deprecated or discouraged) to create an 'aggregate'
builder. The files they see are all files created by hand AND all files
generated by previous builders, if requested.

We now have just a single generated Dart file which collects all of the
parts generated before it.

NOTE: These builders are just an example and in reality a single builder
would be enough to get the same result.