Cleanup part files with a post process builder
There exists a concept of post process builders which are run after
'normal' builders. They don't generate output but require input to work
on.

An example post processor from the build package itself is one that
'cleans up' intermediate files. It doesn't actually delete the files
physically; rather, it just 'deregisters' them as build output.
(According to
https://github.com/dart-lang
/build/issues/3298
it seems not be of much use as of today.)