Add a custom source folder
By default not all files are passed to builders, just the ones that
correspond to package layout conventions, and our custom images/ folder
isn't considered. See
https://github.com/dart-lang/build/blob/master/docs/faq.md#how-can-i-include-additional-sources-in-my-build
for more details.

In our case, we define a custom target with custom sources and enable
only our builder there. We also need to disable our builder from the
$default target because otherwise we would get an overlapping outputs
error - our builder would be applied to both targets. We could have done
it differently in this case as we only have one builder - just change
the sources for the $default target - but in the general case this would
interfere with any other builders that do expect standard sources, or we
would have to configure them explicitly.

Now our custom inputs are discovered and printed out by the builder.