Generate assets to another folder
By default outputs are generated right next to the inputs with simple
extension replacement. To be able to do what we want we need to use
'capture groups'.

We only use one but there can be more and must be named like '{{name}}'
and all capture groups defined in the input must be consumed in the
output - which also means that it is still impossible* to move files in
arbitrary folder structures.

*We can actually do it if we really want by directly accessing the
filesystem and not doing it via the BuildStep abstraction, but this will
definitely prevent the build system from managing the outputs etc. as it
will not know about them. It may be a hacky but viable solution, though.

It may also seem strange that the capture groups are defined in the
configuration of the implementer of the builder and not in the client.
This means that the builder provider decides on the output folder
structure. It is possible to define outputs based on configuration, but
it might break the dependencies and ordering in the build system due to
it being based on static build.yaml files:
https://github.com/dart-lang
/build/issues/3295
In reality, builder team said that as long as the extensions stay the
same all should be good so it is Ok to allow clients to reconfigure
build_extensions. We will actually see that one of their packages allows
overriding it in their own builders so it can't be that bad.

We now commit the assets as well.