Define builder dependencies
As we previously committed the PNG images the asset accessor builder can
find them without issues. However, it would be nice if it could wait for
the image generator to have all the assets in place when it generates
code. For this to happen we can add another target and define a
dependency between them.

We now add another file, images/yellow.b64, and run the build - the
assets/yellow.png file is correctly generated and also included in the
fl_ass.g.dart file.

The build.yaml file now also shows how a second assets folder could be
added (currently commented out).

NOTE: We could add the asset access builder to the $default target.
However, if we change the sources we want to change them only for the
single builder so a custom target is a good option.

NOTE: For me it actually works without defining target dependency but it
is pure coincidence and it could change when more targets and builders
are added. If we are explicit it will never break.

NOTE: About target sources - it turns out they actually extend the
sources _for all targets and all builders within them_ and also set the
default generate_for for the target that specifies them. This is very
confusing and badly documented/explained, unfortunately. As a result, a
strict sources separation among targets is impossible - sources added by
one target are visible to all other builders and they can work on these
sources as well although IMHO they shouldn't even be aware of them. This
is how things work, though, and there are no plans to change it.
https://github.com/dart-lang
/build/issues/3302
But if target sources only add to all sources visible by the builders
why do we explicitly have to include $package$, shouldn't it be included
already? This is somehow treated differently. Lots of questions, so
little documentation.