Implement validator generation
It generates a separate library with extension on the classes it
creates validators for (supports multiple classes per input file).

All lints are disabled because we never know what lints the client
package has enabled so it is virtually impossible to generate lint-free
code for all possible clients.

The code is also formatted by default due to source_gen using dart_style
internally.

The code is generated 'by hand' writing strings; it is also possible to
use the package code_builder but it requires a lot of fiddling with
strings anyway and results in way more code which is even less readable,
at least for this use case. Another way would be to use a template
engine like mustache, but let's keep things simple.