Switch to SharedPartBuilder
Our part can now play nice with parts from other builders and they can
all be combined into a single resulting file.

While it is possible to configure ignored lints etc. in a target for
the combining builder, we decided not to do it here because the
analysis_options.yaml file already excludes all .g.dart files.

As a side-effect of the switch the generator can now be a bit simpler as
it doesn't need to generate import statements. Also using the validator
is easier as there is no need to explicitly import the generated
validator library.