Enable AppCheck to protect the backend
AppCheck needs to be enabled in the Firebase project, and apps using it
need to be registered. Follow
https://firebase.google.com/docs/app-check/flutter/default-providers

NOTE: At this time, only SafetyNet is available for Android (although it
will soon be deprecated) and Device Check for iOS so make sure to
register the correct providers.

Now only registered apps can make calls to APIs which have AppCheck
enabled; all other requests will fail. This is done to protect against
e.g. running bots agains Firestore and eating up all the quota or even
generating exorbitant costs.

NOTE: With AppCheck enforced real devices are needed. It is possible to
configure debug providers but as of now it requires messing around with
native code and I don't feel like doing it.