Add Firebase support
Do this using the new Dart initialization with Firebase CLI and
FlutterFire. Follow
https://firebase.google.com/docs/flutter/setup?platform=android

This will create the Firebase project, register apps and so on for you.

NOTE: The configuration files (google-services.json etc.) are ignored,
it is expected to use a dedicated project (new or existing) for this
branch. This is required, the project won't build otherwise.

NOTE: Whenever it is necessary to choose a region, the code in this
branch assumes 'europe-west3'. This will be important later on, when
Functions are used, so make sure to use it or update the code
accordingly. This is done to reduce latency.