Integrate Cloud Firestore
It will keep FCM tokens per user.

NOTE: minSdkVersion is upgraded to 21 to have multidex support for free,
which is now required due to too many methods for a single DEX.

Firebase CLI is used to manage Firestore indexes, rules, and more in the
future. It can be installed with Homebrew, among others.

Firestore configuration is stored in the firebase/ folder. Issuing
firebase deploy
there deploys everything. This is basically an infrastructure as code
kind of thing.

The rules allow any authenticated user (currently with the anonymous
provider, i.e. anybody running the app) to create a document in the
'users' collection for their 'userId', and allow reading, updating and
deleting this data only by the user themselves.

For the 'firebase deploy' command to work, the CLI needs to know what
project it works with. This is set in firebase/.firebaserc under the
'default' alias. It is currently assumed the id of the Firebase project
is 'firebase-experiments-17', but change accordingly if necessary.