Integrate FCM
It is now possible to receive notifications:
- When the app is in the background, notifications are processed by FCM
  and put in system tray/notification area; tapping the notification
  brings the app to the foreground, starting it if necessary.
- When the app is in the foreground, notifications and custom data are
  available in the callback; FCM does not put them into the system tray
  by default, but this can be changed with priorities if necessary. For
  now our app just logs the message.

NOTE: It is necessary to start the application at least once for
notifications to arrive.

The Firebase console allows sending test notifications using the
composer (but it is very limited in what it can do, e.g. custom data is
not supported).

NOTE: to be able to receive notifications on iOS additional
configuration is needed:
https://firebase.google.com/docs/cloud-messaging/ios/client#upload_your_apns_authentication_key
An Apple developer account is needed for this to work.

In order to be able to send messages to devices their FCM token is
needed. It is kept in Firestore and updated whenever the app is started
or the token changes. This will come in handy in the future.