Implement helper to send FCM messages
It can be invoked with
dart fcm_send.dart <FCM token> [message]

NOTE: For the helper to work a file service-account-key.json is
necessary in the root folder of the repository (it is ignored by Git as
it contains sensitive information). To get one:
1. Go to Firebase project settings.
2. Open the 'Service accounts' tab.
3. In the 'Firebase Admin SDK' section click the 'Generate new private
   key' button.
4. Save the key in the root folder of this repository using the name
   'service-account-key.json'.

Now the application shows a SnackBar when a notification arrives if the
application is in the foreground.

NOTE: It may take a while (even a minute or two) for the notification to
arrive, so have patience.