Handle notifications correctly
Now a SnackBar is shown no matter the app state (foreground, background,
terminated) when a notification arrives (and is clicked if the app is
not in the foreground). It makes use of the built-in functionality for
Flutter to easily detect whether the app was started or resumed by a
notification click.

NOTE: It may take a while (even a minute or two) for the notification to
arrive, so have patience.