Integrate Cloud Functions
Functions is basically a Node.js project with Firebase dependencies. It
is stored in firebase/functions/ and needs to be deployed before the
functions are usable. In the firebase/ folder run

firebase deploy

NOTE: The first time it is necessary to install all the dependencies, so
in firebase/functions/ run

npm install

The example function is a callable function (it means must be actively
invoked), but there are also background functions which will be shown
later.

Again, only verified apps can call Functions. However, it needs to be
explicitly programmed in the functions themselves, which is a bummer.

NOTE: For Function to work the project needs to be upgraded to the Blaze
plan, which means a payment method needs to be defined and the project
can start generating costs. However, due to AppCheck and budget alerts
it is rather unlikely (although still possible).