Send messages
Any user can send messages - they are written to the 'messages'
collection, include some metadata and are guarded by more Firestore
rules. The messages don't appear in the app yet as no reading code is
implemented yet.

NOTE: Make sure to invoke
firebase deploy
in firebase/ for the new rules to be deployed; otherwise it will be
impossible to save new messages. This will also prompt for removal of
the now unnecessary functions.