Use Functions to manage topic subscriptions
Users are subscribed to the 'messages' topic when created and
unsubscribed when deleted. This topic will be used to send out new
messages to all subscribers using FCM.

This is done using 'background' functions that listen to changes in
Firestore documents in the tokens collection.

NOTE: Again, deploy the firebase/ folder for subscription management to
work.