Send out message notifications
If the application is in the background or terminated, new messages
cause notifications to be shown. This works due to a new function that
is triggered whenever a new message is saved and posts its contents to
the 'messages' topic that app instances are subscribed.

NOTE: Again, deploy the firebase/ folder for subscription management to
work.